+++ bot.py
@@ -1,7 +1,4 @@
 import os
-for env in os.environ:
-    print(env, "=", os.environ[env])
-
 from bot import Server
