+++ bot/parser/__init__.py
@@ -1,0 +1,2 @@
+from .text import Text
+from .photo import Photo
