+++ bot/parser/photo.py
@@ -1,0 +1,5 @@
+from .message import Message
+
+class Photo(Message):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
